Revert "Merge pull request #13 from kunimitaiyoh/develop"
This reverts commit 0a4a4efc73a857cd8f77e038da2805dca0c4b809.
