+++ .travis.yml
@@ -1,6 +1,6 @@
 language: node_js
 node_js:
-  - "10"
+  - "8"
 before_script:
 install:
   - npm install
