+++ package.json
@@ -27,32 +27,38 @@
   "homepage": "https://github.com/kunimitaiyoh/kunimitaiyoh.github.io#readme",
   "dependencies": {
     "classnames": "^2.2.5",
-    "preact": "^8.4.2",
-    "preact-router": "^2.6.1",
-    "semantic-ui-css": "^2.4.1"
+    "immutable": "^3.8.2",
+    "js-joda": "^1.8.2",
+    "react": "^16.3.2",
+    "react-dom": "^16.3.3",
+    "semantic-ui-css": "^2.3.1"
   },
   "devDependencies": {
     "@types/classnames": "^2.2.3",
-    "awesome-typescript-loader": "^5.2.1",
-    "css-loader": "^2.1.1",
+    "babel": "^6.23.0",
+    "babel-loader": "^7.1.4",
+    "babel-plugin-jsx-control-statements": "^3.2.8",
+    "babel-plugin-transform-react-jsx": "^6.24.1",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "css-loader": "^0.28.11",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "file-loader": "^1.1.11",
+    "gh-pages": "^1.1.0",
     "glob-all": "^3.1.0",
     "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^3.2.0",
+    "html-webpack-plugin": "^3.0.7",
     "mini-css-extract-plugin": "^0.4.0",
     "purify-css": "^1.2.5",
     "purifycss-webpack": "^0.7.0",
     "sass-loader": "^7.0.1",
     "script-ext-html-webpack-plugin": "^2.0.1",
-    "source-map-loader": "^0.2.4",
     "style-loader": "^0.21.0",
-    "tsconfig-paths-webpack-plugin": "^3.2.0",
-    "typescript": "^3.4.5",
     "uglifyjs-webpack-plugin": "^1.2.5",
     "url-loader": "^1.0.1",
     "webpack": "^4.8.1",
-    "webpack-cli": "^3.3.2",
-    "webpack-dev-server": "^3.3.1"
+    "webpack-cli": "^2.1.3",
+    "webpack-dev-server": "^3.1.11"
   }
 }
