+++ public/index.html
@@ -3,8 +3,6 @@
 <head>
   <meta charset="utf-8" />
   <title>Profile: KUNIMI Taiyoh</title>
-  <link rel="icon" href="/favicon.ico">
-
   <!-- Global site tag (gtag.js) - Google Analytics -->
   <script async src="https://www.googletagmanager.com/gtag/js?id=UA-39499181-3"></script>
   <script>
