+++ src/App.scss
@@ -1,0 +1,4 @@
+.ui.container {
+    margin-top: 3em;
+    margin-bottom: 3em;
+}
