--- src/config.ts
@@ -1,4 +1,0 @@
-declare const WEBPACK_BUILD_DATE: string;
-
-export const BUILD_DATE: string = WEBPACK_BUILD_DATE;
-export const BIRTHDATE = "1990-12-21";
