--- src/views/Profile.scss
@@ -1,13 +1,0 @@
-.ui.container {
-    margin-top: 3em;
-    margin-bottom: 3em;
-}
-
-.period {
-    margin-bottom: 0.5rem;
-}
-
-th {
-    min-width: 10rem;
-    text-align: left;
-}
