--- tsconfig.json
@@ -1,31 +1,0 @@
-{
-    "compilerOptions": {
-      "baseUrl": ".",
-      "rootDirs": ["src"],
-      "module": "esnext",
-      "target": "es6",
-      "moduleResolution": "node",
-      "outDir": "dist",
-      "sourceMap": true,
-      "lib": [
-        "es5",
-        "es6",
-        "es7",
-        "es2018",
-        "dom",
-      ],
-      "jsx": "react",
-      "jsxFactory": "h",
-      "types":  ["preact"],
-      "paths": {
-        "@/*": ["src/*"],
-      },
-      "strict": true,
-      "experimentalDecorators": true
-    },
-    "include": ["./src/**/*"],
-    "exclude": [
-      "dist",
-      "node_modules"
-    ]
-}
